requirements.txt
asgiref==3.5.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
coreapi==2.3.3
coreschema==0.0.4
cryptography==36.0.1
defusedxml==0.7.1
Django==4.0.2
django-filter==21.1
django-templated-mail==1.1.1
djangorestframework==3.13.1
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
drf-extra-fields==3.2.1
flake8==4.0.1
idna==3.3
itypes==1.2.0
Jinja2==3.0.3
MarkupSafe==2.1.0
mccabe==0.6.1
node==0.9.28
npm==0.1.1
oauthlib==3.2.0
odict==1.8.0
optional-django==0.1.0
Pillow==9.0.1
plumber==1.6
psycopg2==2.9.3
psycopg2-binary==2.9.3
pycodestyle==2.8.0
pycparser==2.21
pyflakes==2.4.0
PyJWT==2.3.0
python-dotenv==0.19.2
python3-openid==3.2.0
pytz==2021.3
reportlab==3.6.8
requests==2.27.1
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.2.0
sqlparse==0.4.2
tzdata==2021.5
uritemplate==4.1.1
urllib3==1.26.8
zope.component==5.0.1
zope.deprecation==4.4.0
zope.event==4.5.0
zope.hookable==5.1.0
zope.interface==5.4.0
zope.lifecycleevent==4.3
